requirements.txt
Flask
Flask-SQLAlchemy
PyJWT
Werkzeug
python-dotenv
gunicorn
psycopg2-binary